LCD_5110.* Updated to version 1.2. Fixed function write_xy().
brd_driver_hw_03.* Version value is now generated as data and time when
  preprocessor is run. So it generate always unique number with date
  stamp, so now can be easily tracked which firmware is actual.

2DO: Add support for different bit resolution
     Support DPS, left and right justify
     Tidy error, warning and info messages in brd_driver_hw_03.h
     Bootloader

+++ FW/src/LCD_5110.h
@@ -8,9 +8,9 @@
  * Please include correct HAL driver files.
  *
  * Created:  04.07.2012\n
- * Modified: 06.05.2014
+ * Modified: 11.09.2014
  *
- * \version 1.1
+ * \version 1.2
  * \author Martin Stejskal
  */
 
@@ -185,7 +185,7 @@ e_lcd_5110_status LCD_5110_init(void);
  *  @return LCD_5110_OK (0) if all OK
  */
  e_lcd_5110_status LCD_5110_write_xy(const char * p_txt_to_show,
-                                     int x_position, int y_line);
+                                     uint8_t x_position, uint8_t y_line);
 
  /**
   * \brief Clear selected line and set cursor to X position 0
